Correct AbstractState_set_binary_interaction_double in shared library (#1117)
* Correct AbstractState_set_binary_interaction_double

Convert `i` and `j` to `size_t` when calling `AbstractState_set_binary_interaction_double`

* Addapt Julia wrapper to c16a4b6

Now `AbstractState_set_binary_interaction_double` needs `long` for `i` and `j`

+++ src/CoolPropLib.cpp
@@ -653,7 +653,7 @@ EXPORT_CODE void CONVENTION AbstractState_set_binary_interaction_double(const lo
     *errcode = 0;
     try {
         shared_ptr<CoolProp::AbstractState> &AS = handle_manager.get(handle);
-        AS->set_binary_interaction_double(i, j, parameter, value);
+        AS->set_binary_interaction_double(static_cast<std::size_t>(i), static_cast<std::size_t>(j), parameter, value);
     }
     catch (CoolProp::HandleError &e) {
         std::string errmsg = std::string("HandleError: ") + e.what();
